--- app/src/main/java/com/example/foodproject/Activity/BaseActivity.java
@@ -1,24 +1,0 @@
-package com.example.foodproject.Activity;
-
-import android.os.Bundle;
-import android.view.Window;
-import android.view.WindowManager;
-import androidx.appcompat.app.AppCompatActivity;
-import com.example.foodproject.R;
-
-import com.google.firebase.database.FirebaseDatabase;
-
-public class BaseActivity extends AppCompatActivity {
-    FirebaseDatabase database;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-        database = FirebaseDatabase.getInstance();
-        setContentView(R.layout.activity_base);
-        Window w = getWindow();
-        w.setFlags(WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
-                WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS);
-    }
-}
