+++ app/src/main/java/com/example/foodproject/Activity/Helper/ChangeNumberItemsListener.java
@@ -1,4 +1,4 @@
-package com.example.foodproject.Helper;
+package com.example.foodproject.Activity.Helper;
 
 public interface ChangeNumberItemsListener {
     void change();
