+++ app/src/main/java/com/example/foodproject/Activity/Helper/TinyDB.java
@@ -19,7 +19,7 @@
  *  and unicode 2017 that are used for separating the items in a list.
  */
 
-package com.example.foodproject.Helper;
+package com.example.foodproject.Activity.Helper;
 
 import android.content.Context;
 import android.content.SharedPreferences;
@@ -32,7 +32,7 @@ import android.text.TextUtils;
 import android.util.Log;
 
 
-import com.example.foodproject.Domain.Foods;
+import com.example.foodproject.Activity.Domain.Foods;
 import com.google.gson.Gson;
 
 import java.io.File;
