+++ app/src/main/res/layout/activity_detail.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    tools:context=".Activity.DetailActivity">
+    tools:context=".Activity.Activity.DetailActivity">
 
     <View
         android:id="@+id/view"
