+++ app/src/main/res/layout/activity_list_food.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/yellow"
-    tools:context=".Activity.ListFoodActivity">
+    tools:context=".Activity.Activity.ListFoodActivity">
 
     <ScrollView
         android:id="@+id/scrollView4"
